+++ app/src/js/components/Collections/granules.js
@@ -26,7 +26,6 @@ import { strings } from '../locale';
 import { workflowOptionNames } from '../../selectors';
 import ListFilters from '../ListActions/ListFilters';
 import Breadcrumbs from '../Breadcrumbs/Breadcrumbs';
-import pageSizeOptions from '../../utils/page-size';
 import { getPersistentQueryParams } from '../../utils/url-helper';
 
 const CollectionGranules = ({
@@ -172,6 +171,8 @@ const CollectionGranules = ({
           rowId="granuleId"
           sortId="timestamp"
           tableColumns={tableColumns}
+          filterAction={filterGranules}
+          filterClear={clearGranulesFilter}
         >
           <ListFilters>
             <Search
@@ -193,16 +194,6 @@ const CollectionGranules = ({
                 }}
               />
             )}
-            <Dropdown
-              options={pageSizeOptions}
-              action={filterGranules}
-              clear={clearGranulesFilter}
-              label="Results Per Page"
-              paramKey="limit"
-              inputProps={{
-                placeholder: 'Results Per Page',
-              }}
-            />
           </ListFilters>
         </List>
       </section>