+++ app/src/js/components/Collections/overview.js
@@ -21,7 +21,6 @@ import {
   getCollectionId,
   lastUpdated,
 } from '../../utils/format';
-import pageSizeOptions from '../../utils/page-size';
 import statusOptions from '../../utils/status';
 import { getPersistentQueryParams, historyPushWithQueryParams } from '../../utils/url-helper';
 import {
@@ -284,6 +283,8 @@ class CollectionOverview extends React.Component {
             bulkActions={this.generateBulkActions()}
             rowId="granuleId"
             sortId="timestamp"
+            filterAction={filterGranules}
+            filterClear={clearGranulesFilter}
           >
             <ListFilters>
               <Search
@@ -301,16 +302,6 @@ class CollectionOverview extends React.Component {
                 label="Status"
                 inputProps={{
                   placeholder: 'All',
-                }}
-              />
-              <Dropdown
-                options={pageSizeOptions}
-                action={filterGranules}
-                clear={clearGranulesFilter}
-                paramKey="limit"
-                label={'Results Per Page'}
-                inputProps={{
-                  placeholder: 'Results Per Page',
                 }}
               />
             </ListFilters>