+++ app/src/js/components/Executions/overview.js
@@ -16,7 +16,6 @@ import {
 import { tally, lastUpdated } from '../../utils/format';
 import { workflowOptions as workflowSelectOptions } from '../../selectors';
 import statusOptions from '../../utils/status';
-import pageSizeOptions from '../../utils/page-size';
 import List from '../Table/Table';
 import Dropdown from '../DropDown/dropdown';
 import Search from '../Search/search';
@@ -88,6 +87,8 @@ class ExecutionOverview extends React.Component {
             query={{ ...queryParams }}
             rowId='name'
             sortId='createdAt'
+            filterAction={filterExecutions}
+            filterClear={clearExecutionsFilter}
           >
             <ListFilters>
               <Dropdown
@@ -131,17 +132,6 @@ class ExecutionOverview extends React.Component {
                 paramKey={'asyncOperationId'}
                 label={'Async Operation ID'}
                 placeholder="Search"
-              />
-
-              <Dropdown
-                options={pageSizeOptions}
-                action={filterExecutions}
-                clear={clearExecutionsFilter}
-                paramKey={'limit'}
-                label={'Results Per Page'}
-                inputProps={{
-                  placeholder: 'Results Per Page',
-                }}
               />
             </ListFilters>
           </List>