+++ app/src/js/components/ReconciliationReports/list.js
@@ -105,6 +105,8 @@ class ReconciliationReportList extends React.Component {
             bulkActions={this.generateBulkActions()}
             rowId='name'
             sortId='createdAt'
+            filterAction={filterReconciliationReports}
+            filterClear={clearReconciliationReportsFilter}
           >
             <ListFilters>
               <Search